# [PERF]: dont read parquet metadata multiple times