related to https://github.com/Eventual-Inc/Daft/issues/2257


Notes for reviewer:

this kind of feels like a "brute force" approach as I just added `parquet_metadata` field to `AnonymousDataFile`, but I couldn't find a better way of doing it without sufficient refactoring. If there are better alternatives, I'm all ears 😄 

I also didn't really like that I had to pass it down to all of the different `*_read_parquet_*` functions, but once again, I wanted to avoid unnecessary refactoring. 



Some local tests using the file mentioned in the issue show over 100% increase. 

- main: 25sec 782ms
- feature branch: 11sec 641ms